feat: Refactor state management and view models
This commit introduces several changes to the app's state management and view models, including:

-   Created `SavedRecipesState` to manage the state of saved recipes, `RecipeSearchState` to manage the state of recipe searches, and `RecipeDetailState` to manage the state of a specific recipe's details.
-   Moved the definition of `SavedRecipesState`, `RecipeSearchState` and `RecipeDetailState` to their corresponding package.
-   Renamed `RecipeViewModel` to `SavedRecipeViewModel`, `SearchViewModel` to `RecentlySearchViewModel`, and moved them into their corresponding package.
-   Renamed `MainActivity.kt` `RecipeViewModel` import to `SavedRecipeViewModel` import.
-   Created `SavedRecipes.kt` which will be the screen for saved recipes and it uses the `SavedRecipesState` and `SavedRecipeViewModel`
