+++ app/src/main/java/com/surivalcoding/composerecipeapp/presentation/recently_search_recipe/RecentlySearchState.kt
@@ -1,0 +1,9 @@
+package com.surivalcoding.composerecipeapp.presentation.saved_recipes.view_model
+
+import Recipe
+
+data class RecipeSearchState(
+    val keyword: String = "",
+    val recipeList: List<Recipe> = emptyList(),
+    val isLoading: Boolean = false,
+)
