+++ app/src/main/java/com/surivalcoding/composerecipeapp/presentation/saved_recipes/SavedRecipesState.kt
@@ -1,0 +1,8 @@
+package com.surivalcoding.composerecipeapp.presentation.saved_recipes
+
+import Recipe
+
+data class SavedRecipesState(
+    val recipes: List<Recipe> = emptyList(),
+    val isLoading: Boolean = false
+)
